+++ src/FarNet.Redis/Commands/UseLock.cs
@@ -1,8 +1,5 @@
 ﻿using StackExchange.Redis;
-using System;
 using System.Diagnostics;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace FarNet.Redis.Commands;