+++ src/PS.FarNet.Redis/Abc.cs
@@ -1,7 +1,5 @@
 ﻿using StackExchange.Redis;
-using System;
 using System.Collections;
-using System.Collections.Generic;
 using System.Management.Automation;
 
 namespace PS.FarNet.Redis;