+++ src/PS.FarNet.Redis/GetAnyCommand.cs
@@ -1,6 +1,5 @@
 ﻿using StackExchange.Redis;
 using System.Collections;
-using System.Collections.Generic;
 using System.Management.Automation;
 
 namespace PS.FarNet.Redis;