+++ src/PS.FarNet.Redis/GetSetCommand.cs
@@ -1,5 +1,4 @@
 ﻿using StackExchange.Redis;
-using System.Collections.Generic;
 using System.Management.Automation;
 
 namespace PS.FarNet.Redis;