+++ src/PS.FarNet.Redis/SetHashCommand.cs
@@ -1,5 +1,4 @@
 ﻿using StackExchange.Redis;
-using System;
 using System.Collections;
 using System.Management.Automation;